Fix for when running with root (vagrant box)
When running as root during the provision step with vagrant, the home
directory is set as root instead of the vagrant user, causing issues
such as where the demo_rails_app will go as well as during the check for
any previous rails apps.

I now put in a check if this will be run in a vagrant box via a command
line argument which will hard set the directory and ignore if the folder
already exists since sharing the folder creates the folder when running
under vagrant.